More about the pie game
I've been learning Python for a few weeks now and I wanted to do something to challenge myself. So, I create a fun game by creating a while loop inside of a function. I created 2 functions, one function "my_program" will ask for your name and then read it back to you and then tell the rules of the game. 
I create a second function "chances" when the function is called, the while loop inside of it will commence and from there the user will have 6 attempts to try to guess my favorite flavor of pie. 
I thought it would be cleaner and a little more challenging to combine the function and while loop. This took me a few hours to create and because I'm still learning and to enjoy challenges, as a condition of the if statement, if the user guesses the right flavor, regardless if the guess can be in all caps, lower case or begin with an uppercase letter, it will still be correct. This particular piece was challenging and it took some trial and error and practice to get it correct. Initially, the the loop stopped even when the guess was incorrect. I still have so much to learn still, but this felt so satisfying to finish and see that it works how it  should. I now see what I've been reading how programmers feel when they code and see their finished product after having trial and error and some headbanging moments. It's so worth it to work through it. I wanted to break out of "tutorial hell" and start to create more things and I learned a lot about Python and myself in just creating this game.

I gladly welcome any feedback and happy coding, friends! :)